Use the actual put_bulk request for test accuracy instead of performance

+++ ds3_java_cli/src/main/java/com/spectralogic/ds3cli/command/Performance.java
@@ -73,7 +73,7 @@ public class Performance extends CliCommand<DefaultResult> {
     public DefaultResult call() throws Exception {
         final Ds3ClientHelpers helpers = Ds3ClientHelpers.wrap(getClient());
         final int numberOfFiles = Integer.valueOf(this.numberOfFiles);
-        final long sizeOfFiles = Integer.valueOf(this.sizeOfFiles);
+        final long sizeOfFiles = Long.valueOf(this.sizeOfFiles);
 
         try {
             try {
